requirements.txt
absl-py==2.3.1
addict==2.4.0
aiomysql==0.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
attrs==25.4.0
basicsr==1.4.2
bcrypt==3.2.0
beautifulsoup4==4.14.2
blinker==1.9.0
Brlapi==0.8.3
certifi==2025.11.12
cffi==2.0.0
chardet==4.0.0
click==8.3.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
curl_cffi==0.13.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.3.0
facexlib==0.3.0
fastapi==0.121.3
fasteners==0.14.1
filelock==3.20.0
filterpy==1.4.5
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
folder-color==0.2.0
fonttools==4.60.1
fsspec==2025.10.0
future==0.18.2
gfpgan==1.3.8
greenlet==3.2.4
grpcio==1.76.0
gyp==0.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperframe==6.1.0
icoextract==0.1.2
idna==3.3
ImageIO==2.37.2
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.7.0
mysqlclient==2.2.7
netifaces==0.11.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
olefile==0.46
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
paramiko==2.9.3
pefile==2021.9.3
pexpect==4.8.0
pillow==12.0.0
platformdirs==4.5.0
playwright==1.56.0
postgrest==2.24.0
propcache==0.4.1
protobuf==6.33.1
psutil==5.9.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyaes==1.6.1
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
pyelftools==0.27
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyMySQL==1.1.2
PyNaCl==1.5.0
pyparsing==3.2.5
pyperclip==1.11.0
pyRFC3339==1.1
Pyrogram==2.0.106
PySocks==1.7.1
python-apt===2.4.0-ubuntu4-zorin1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==8.2
realtime==2.24.0
reportlab==3.6.8
requests==2.25.1
scikit-image==0.25.2
scipy==1.15.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.38.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.50.0
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
TgCrypto==1.2.5
tifffile==2025.5.10
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.38.0
wadllib==1.3.6
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.1
xdg==5
xkit==0.0.0
yapf==0.43.0
yarl==1.22.0
zipp==1.0.0
zorin-appearance==5.3
zorin-connect==1.2
zorin-exec-guard==1.5

rembg==2.0.57

onnxruntime==1.17.1
